tiny-universe/TinyRAG/requirements.txt
openai
zhipuai
numpy
python-dotenv
torch
torchvision
torchaudio
transformers
tqdm
PyPDF2
markdown
html2text
tiktoken
beautifulsoup4
sentencepiece
modelscope